Start refactoring methods
