+++ code/api/imports/data/collection/methods/User/fetchOneUser.js
@@ -1,0 +1,3 @@
+import { Meteor } from 'meteor/meteor'
+
+export const fetchOneUser = _id => Meteor.users.findOne({ _id })
