+++ code/api/imports/data/collection/methods/User/followerUser.js
@@ -1,0 +1,4 @@
+import { follow } from '../../../../lib/Follower/FollowerMethods'
+import { userCollection } from '../../UserCollection'
+
+export const followUser = follow(userCollection)
